+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2023-03-02
+
+### Added
+
+- Add Client#transcribe endpoint - Whisper transcribtion over the wire!
+- Add Client#translate endpoint - Whisper translation over the wire!
 ## [3.4.0] - 2023-03-01
 
 ### Added